+++ src/components/App.vue
@@ -10,6 +10,7 @@
         />
       </keep-alive>
     </router-view>
+    <all-dialog :is-engines-ready="isEnginesReady" />
   </error-boundary>
 </template>
 
@@ -21,6 +22,13 @@ import { EngineId } from "@/type/preload";
 import ErrorBoundary from "@/components/ErrorBoundary.vue";
 import { useStore } from "@/store";
 import { useHotkeyManager } from "@/plugins/hotkeyPlugin";
+import {
+  DEFAULT_BEATS,
+  DEFAULT_BEAT_TYPE,
+  DEFAULT_BPM,
+  DEFAULT_TPQN,
+} from "@/sing/storeHelper";
+import AllDialog from "@/components/Dialog/AllDialog.vue";
 
 const store = useStore();
 const route = useRoute();
@@ -107,7 +115,36 @@ onMounted(async () => {
   isEnginesReady.value = true;
 });
 
-// TODO: ダイアログ周りをEditorHomeから移動する
+onMounted(async () => {
+  await store.dispatch("SET_SCORE", {
+    score: {
+      tpqn: DEFAULT_TPQN,
+      tempos: [
+        {
+          position: 0,
+          bpm: DEFAULT_BPM,
+        },
+      ],
+      timeSignatures: [
+        {
+          measureNumber: 1,
+          beats: DEFAULT_BEATS,
+          beatType: DEFAULT_BEAT_TYPE,
+        },
+      ],
+      notes: [],
+    },
+  });
+
+  await store.dispatch("SET_VOLUME", { volume: 0.6 });
+  await store.dispatch("SET_PLAYHEAD_POSITION", { position: 0 });
+  await store.dispatch("SET_LEFT_LOCATOR_POSITION", {
+    position: 0,
+  });
+  await store.dispatch("SET_RIGHT_LOCATOR_POSITION", {
+    position: 480 * 4 * 16,
+  });
+});
 
 // TODO: エンジン起動状態周りの処理と表示をEditorHomeから移動する
 </script>
