+++ src/components/Sing/SingEditor.vue
@@ -25,16 +25,10 @@
 </template>
 
 <script setup lang="ts">
-import { computed, onMounted, watch } from "vue";
+import { computed, watch } from "vue";
 import MenuBar from "./MenuBar.vue";
 import ToolBar from "./ToolBar.vue";
 import ScoreSequencer from "./ScoreSequencer.vue";
-import {
-  DEFAULT_BEATS,
-  DEFAULT_BEAT_TYPE,
-  DEFAULT_BPM,
-  DEFAULT_TPQN,
-} from "@/sing/storeHelper";
 import EngineStartupOverlay from "@/components/EngineStartupOverlay.vue";
 import { useStore } from "@/store";
 
@@ -62,39 +56,6 @@ const nowAudioExporting = computed(() => {
 const cancelExport = () => {
   store.dispatch("CANCEL_AUDIO_EXPORT");
 };
-
-// 歌声合成エディターの初期化
-onMounted(async () => {
-  await store.dispatch("SET_SCORE", {
-    score: {
-      tpqn: DEFAULT_TPQN,
-      tempos: [
-        {
-          position: 0,
-          bpm: DEFAULT_BPM,
-        },
-      ],
-      timeSignatures: [
-        {
-          measureNumber: 1,
-          beats: DEFAULT_BEATS,
-          beatType: DEFAULT_BEAT_TYPE,
-        },
-      ],
-      notes: [],
-    },
-  });
-
-  await store.dispatch("SET_VOLUME", { volume: 0.6 });
-  await store.dispatch("SET_PLAYHEAD_POSITION", { position: 0 });
-  await store.dispatch("SET_LEFT_LOCATOR_POSITION", {
-    position: 0,
-  });
-  await store.dispatch("SET_RIGHT_LOCATOR_POSITION", {
-    position: 480 * 4 * 16,
-  });
-  return {};
-});
 
 // エンジン初期化後の処理
 const unwatchIsEnginesReady = watch(
